Move create_rti_server to socket_common.c
